* moved a few methods into the toolkit * cleaned up lots of imports * augmented checkpoint yml comments
