+++ great_expectations/cli/tap.py
@@ -3,14 +3,13 @@ import sys
 
 import click
 
-from great_expectations.cli.datasource import get_batch_kwargs, \
-    select_datasource
+import great_expectations.cli.toolkit as toolkit
+from great_expectations.cli.datasource import get_batch_kwargs
 from great_expectations.cli.mark import Mark as mark
-from great_expectations.cli.util import load_data_context_with_error_handling, \
-    cli_message
-from great_expectations.cli.toolkit import load_expectation_suite
-from great_expectations.core.usage_statistics.usage_statistics import \
-    send_usage_message
+from great_expectations.cli.util import cli_message
+from great_expectations.core.usage_statistics.usage_statistics import (
+    send_usage_message,
+)
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.util import lint_code
 
@@ -42,12 +41,12 @@ def tap_new(suite, tap_filename, directory, datasource=None):
 
 
 def _tap_new(suite, tap_filename, directory, usage_event, datasource=None):
-    context = load_data_context_with_error_handling(directory)
+    context = toolkit.load_data_context_with_error_handling(directory)
     try:
         _validate_tap_filename(tap_filename)
         context_directory = context.root_directory
         datasource = _get_datasource(context, datasource)
-        suite = load_expectation_suite(context, suite, usage_event)
+        suite = toolkit.load_expectation_suite(context, suite, usage_event)
         _, _, _, batch_kwargs = get_batch_kwargs(context, datasource.name)
 
         tap_filename = _write_tap_file_to_disk(
@@ -82,7 +81,7 @@ def _validate_tap_filename(tap_filename):
 
 
 def _get_datasource(context, datasource):
-    datasource = select_datasource(context, datasource_name=datasource)
+    datasource = toolkit.select_datasource(context, datasource_name=datasource)
     if not datasource:
         cli_message("<red>No datasources found in the context.</red>")
         sys.exit(1)
